+++ pylabrobot/liquid_handling/liquid_handler.py
@@ -1014,10 +1014,9 @@ class LiquidHandler(Machine):
 
     # liquid(s) for each channel. If volume tracking is disabled, use None as the liquid.
     if does_volume_tracking():
-      liquids = [
-        c.get_tip().tracker.get_liquids(top_volume=vol)
-        for c, vol in zip(self.head.values(), vols)
-      ]
+      channels = [self.head[channel] for channel in use_channels]
+      liquids = [c.get_tip().tracker.get_liquids(top_volume=vol)
+                 for c, vol in zip(channels, vols)]
     else:
       liquids = [[(None, vol)] for vol in vols]